Actually follow sensible coding conventions
