+++ Peisik.Compiler/Compiler/Optimizing/Expression.cs
@@ -321,159 +321,4 @@ namespace Polsys.Peisik.Compiler.Optimizing
                     paramCount.ToString(), maxParams.ToString());
         }
     }
-
-    /// <summary>
-    /// Represents a compile-time constant.
-    /// </summary>
-    internal class ConstantExpression : Expression
-    {
-        public object Value;
-
-        public ConstantExpression(object value, OptimizingCompiler compiler, LocalVariable store = null, TokenPosition position = default)
-        {
-            // Store the type
-            switch (value)
-            {
-                case bool b:
-                    Type = PrimitiveType.Bool;
-                    break;
-                case long l:
-                    Type = PrimitiveType.Int;
-                    break;
-                case double d:
-                    Type = PrimitiveType.Real;
-                    break;
-                default:
-                    throw new ArgumentException("Unknown constant type");
-            }
-
-            Value = value;
-            SetStore(store, compiler, position);
-        }
-
-        public ConstantExpression(LiteralSyntax expression, OptimizingCompiler compiler, LocalVariable store = null)
-            : this(expression.Value, compiler, store, expression.Position)
-        {
-        }
-
-        public override Expression Fold(OptimizingCompiler compiler)
-        {
-            return this;
-        }
-
-        protected override void SetStore(LocalVariable newStore, OptimizingCompiler compiler, TokenPosition position = default)
-        {
-            base.SetStore(newStore, compiler, position);
-        }
-    }
-
-    /// <summary>
-    /// Represents a function call.
-    /// The call target is always a user-defined function.
-    /// </summary>
-    internal class FunctionCallExpression : Expression
-    {
-        public Function Callee { get; private set; }
-        public List<Expression> Parameters { get; private set; }
-        public bool DiscardResult { get; private set; }
-
-        public FunctionCallExpression(Function callee, List<Expression> parameters,
-            bool discardResult, OptimizingCompiler compiler)
-        {
-            Callee = callee;
-            Parameters = parameters;
-            DiscardResult = discardResult;
-            Type = callee.ResultValue.Type;
-        }
-
-        public override Expression Fold(OptimizingCompiler compiler)
-        {
-            // CODE SMELL: Immutability violation
-            for (var i = 0; i < Parameters.Count; i++)
-                Parameters[i] = Parameters[i].Fold(compiler);
-
-            return this;
-        }
-
-        protected override void SetStore(LocalVariable newStore, OptimizingCompiler compiler, TokenPosition position = default)
-        {
-            if (DiscardResult && newStore != null)
-                throw new InvalidOperationException("Cannot set Store when DiscardResult is true");
-
-            base.SetStore(newStore, compiler, position);
-        }
-    }
-
-    /// <summary>
-    /// Represents a local variable load.
-    /// </summary>
-    internal class LocalLoadExpression : Expression
-    {
-        public LocalVariable Local;
-
-        public LocalLoadExpression(LocalVariable local, OptimizingCompiler compiler, LocalVariable store = null)
-        {
-            Local = local;
-            Local.UseCount++;
-            Type = local.Type;
-
-            SetStore(store, compiler);
-        }
-
-        public override Expression Fold(OptimizingCompiler compiler)
-        {
-            return this;
-        }
-
-        internal override Expression TryInlineLocalAssignment(Expression assignment)
-        {
-            if (Local == assignment.Store)
-            {
-                return assignment;
-            }
-            return null;
-        }
-
-        protected override void SetStore(LocalVariable newStore, OptimizingCompiler compiler, TokenPosition position = default)
-        {
-            base.SetStore(newStore, compiler, position);
-        }
-    }
-
-    /// <summary>
-    /// Represents a point where the function will return to its caller.
-    /// </summary>
-    /// <remarks>
-    /// If the containing function is inlined, this will become a local store.
-    /// The variables are renamed so that the result variable becomes a local in the inlining function.
-    /// </remarks>
-    internal class ReturnExpression : Expression
-    {
-        /// <summary>
-        /// The return value, which may be null.
-        /// </summary>
-        public Expression Value;
-
-        public ReturnExpression(Expression returnValue)
-        {
-            Value = returnValue;
-        }
-
-        public override Expression Fold(OptimizingCompiler compiler)
-        {
-            if (Value != null)
-                return new ReturnExpression(Value.Fold(compiler));
-            else
-                return this;
-        }
-
-        internal override Expression TryInlineLocalAssignment(Expression assignment)
-        {
-            if (Value is LocalLoadExpression load && load.Local == assignment.Store)
-            {
-                return new ReturnExpression(assignment);
-            }
-            return null;
-        }
-    }
 }
